requirements.txt
torch==2.10.0
torchvision==0.25.0
opencv-python==4.13.0.88
matplotlib==3.10.8
streamlit==1.55.0
pillow==12.1.1
numpy==2.4.3
lxml==6.0.2